requirements.txt
playwright>=1.44.0
requests>=2.31.0
openai>=1.40.0
pydantic>=2.0
tenacity>=8.2
python-dotenv>=1.0.0
tqdm>=4.0
